pyproject.toml
[tool.poetry]
name = "opti"
version = "0.1.0"
description = ""
authors = ["mak.pro.devel <mak.pro.devel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.110.1"}
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
requests = "^2.31.0"
uvicorn = "^0.29.0"
authlib = "^1.3.0"
redis = "^4.0.0"
asyncpg = "^0.29.0"
websockets = "^12.0"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
loguru = "^0.7.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = [
  ".", "opti",
]
asyncio_mode="auto"